requirements.txt
fastapi
uvicorn
google-generativeai
pinecone
pandas
openpyxl
python-multipart
python-dotenv
requests
psycopg2-binary
sqlalchemy
